fixed: 部分文件夹结构错误 todo: log创建位置不对; header还可以简化
